+++ tests/ModularArithmeticSpec.hs
@@ -20,7 +20,3 @@ spec = describe "modular arithmetic" $ do
       keccak256bs 64 4637928374822348932 `shouldBe` 31913459109244394690918113775821069766218386171344476352830309015864679314372
       keccak256bs 128 86734239273823482392374839238192 `shouldBe` 73259886156781299012365827306350972332057929114712718446011773447329916800282
       keccak256bs 256 2939802230983298498274024970323894828329382938283938293283 `shouldBe` 112308053996183008364732118192044491261660825982790234118109580320965363105885
-  describe "poseidon-hash" $ do
-    it "computes the poseidon hash for a given list of integers" $ do
-      poseidon [1, 2] `shouldBe` 12717992376338182279477285556390228582603857817939167998284243525425604090033
-      poseidon [999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999, 88888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888] `shouldBe` 1069652563792426660145833190237336045845917811967528486535167837537743676741
