+++ tests/pact/poseidon-hash.repl
@@ -2,3 +2,4 @@
 
 (expect "1 2 should equal 12717992376338182279477285556390228582603857817939167998284243525425604090033" 12717992376338182279477285556390228582603857817939167998284243525425604090033 (poseidon-hash 1 2))
 (expect "999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999 88888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888 should equal 1069652563792426660145833190237336045845917811967528486535167837537743676741" 1069652563792426660145833190237336045845917811967528486535167837537743676741 (poseidon-hash 999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999 88888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888888))
+
